+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/base/plugins/Plugin.kt
@@ -2,12 +2,14 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins
 
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.FinalConfig
 import com.pengrad.telegrambot.TelegramBot
+import java.util.logging.Level
 import java.util.logging.Logger
 
-typealias PluginVersion = Long
 typealias PluginName = String
 
-val commonLogger = Logger.getLogger("common")
+val commonLogger = Logger.getLogger("common").also {
+    it.level = Level.FINER
+}
 
 @Deprecated(
     "This variable was deprecated for the reason that it is useless in context of commonarchitecture",